lithium.0.2.53 - logo clouds [1]

+++ jni/Geometry.h
@@ -23,9 +23,9 @@ inline float Deg(float Radians) { return Radians * (180.0f / PI); };
 inline float sqr(float x) { return x * x; };
 
 // Calculates random 0 < x < 1.
-inline float frand(float x = 1.0f) { return (float)rand() / RAND_MAX * x; };
+inline float frand(float x = 1.0f) { return (float)rand() / (float)RAND_MAX * x; };
 
-// Methods to return a random value between ranges.
+// Random value between ranges.
 inline float frandRange(float min, float max) { return min + (((float)rand() / (float)RAND_MAX) * (max - min)); };
 
 //-----------------------------------------------------------------------------
